Refactor: separated DesktopPopup from MarkerPopup (#100)
